[Update: Scrapped videos and Serialized them]
The code is not so nice now btw. I want to scrape the videos and images and maybe text only also and maybe convert them to screenshots type maybe. I have used too many maybes mannnn. Also basic idea is to scrape and I don't have any idea for their storage. Cloud storage maybe insufficient in the long run and I will see if it is possible to directly patch up the links. I am confused, whether I should switch to python ffmpeg or use rust's ffmpeg gstreamer because the documentation in rust for two libraries is shit(I know, surprising it is.